+++ automerge/src/transaction.rs
@@ -1,15 +1,14 @@
 mod commit;
 mod inner;
+mod manual_transaction;
 mod result;
 mod transactable;
-mod transaction;
 
+pub use self::commit::CommitOptions;
+pub use self::transactable::Transactable;
 pub(crate) use inner::TransactionInner;
+pub use manual_transaction::Transaction;
 pub use result::TransactionFailure;
 pub use result::TransactionSuccess;
 
 pub type TransactionResult<O, E> = Result<TransactionSuccess<O>, TransactionFailure<E>>;
-
-pub use self::commit::CommitOptions;
-pub use self::transactable::Transactable;
-pub use transaction::Transaction;
